backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-jose[cryptography]==3.5.0
sniffio==1.3.1
starlette==0.47.2
supabase==2.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
python-dotenv==1.0.0
python-multipart==0.0.6
